feat(apiv2): reorganize controllers, add stats to meta endpoint
