+++ PluralKit.API/Controllers/v2/DiscordControllerV2.cs
@@ -5,6 +5,8 @@ using Microsoft.AspNetCore.Mvc;
 
 using Newtonsoft.Json.Linq;
 
+using NodaTime;
+
 using PluralKit.Core;
 
 namespace PluralKit.API
@@ -120,6 +122,26 @@ namespace PluralKit.API
             return Ok(newSettings.ToJson());
         }
 
+        [HttpGet("messages/{messageId}")]
+        public async Task<ActionResult<MessageReturn>> MessageGet(ulong messageId)
+        {
+            var msg = await _db.Execute(c => _repo.GetMessage(c, messageId));
+            if (msg == null)
+                throw Errors.MessageNotFound;
 
+            var ctx = this.ContextFor(msg.System);
+
+            // todo: don't rely on v1 stuff
+            return new MessageReturn
+            {
+                Timestamp = Instant.FromUnixTimeMilliseconds((long)(msg.Message.Mid >> 22) + 1420070400000),
+                Id = msg.Message.Mid.ToString(),
+                Channel = msg.Message.Channel.ToString(),
+                Sender = msg.Message.Sender.ToString(),
+                System = msg.System.ToJson(ctx, v: APIVersion.V2),
+                Member = msg.Member.ToJson(ctx, v: APIVersion.V2),
+                Original = msg.Message.OriginalMid?.ToString()
+            };
+        }
     }
 }
