+++ PluralKit.Core/Database/Repository/ModelRepository.Shards.cs
@@ -9,8 +9,8 @@ namespace PluralKit.Core
 {
     public partial class ModelRepository
     {
-        public Task<IEnumerable<PKShardInfo>> GetShards(IPKConnection conn) =>
-            conn.QueryAsync<PKShardInfo>("select * from shards order by id");
+        public Task<IEnumerable<PKShardInfo>> GetShards() =>
+            _db.Execute(conn => conn.QueryAsync<PKShardInfo>("select * from shards order by id"));
 
         public Task SetShardStatus(IPKConnection conn, int shard, PKShardInfo.ShardStatus status) =>
             conn.ExecuteAsync(
